+++ crates/hayagriva-cli/Cargo.toml
@@ -22,3 +22,4 @@ biblatex = { workspace = true, optional = true }
 
 [features]
 default = ["biblatex"]
+biblatex = ["hayagriva-format/biblatex", "dep:biblatex"]
