+++ crates/hayagriva-cli/src/main.rs
@@ -10,6 +10,8 @@ use citationberg::{
 };
 use clap::builder::PossibleValue;
 use clap::{Arg, ArgAction, Command, ValueEnum, crate_version};
+#[cfg(feature = "biblatex")]
+use hayagriva_format::io::from_biblatex_str;
 use strum::VariantNames;
 
 use hayagriva_archive::{ArchivedStyle, locales};
@@ -233,7 +235,7 @@ fn main() {
         match format {
             Format::Yaml => io::from_yaml_str(&input).unwrap(),
             #[cfg(feature = "biblatex")]
-            Format::Biblatex | Format::Bibtex => io::from_biblatex_str(&input).unwrap(),
+            Format::Biblatex | Format::Bibtex => from_biblatex_str(&input).unwrap(),
         }
     };
 
@@ -484,4 +486,490 @@ fn retrieve_assets<'a>(
     };
 
     (style, locales, locale)
+}
+
+#[cfg(test)]
+mod tests {
+    use std::{borrow::Cow, fs, path::Path, sync::Arc};
+
+    use citationberg::{
+        IndependentStyle, LocaleCode, LocaleFile,
+        taxonomy::{Locator, NumberVariable},
+    };
+    use hayagriva_core::{EntryLike, MaybeTyped, Numeric, NumericValue};
+    use hayagriva_csl::{
+        BibliographyDriver, BibliographyRequest, BufWriteFormat, CitationItem,
+        CitationRequest, CitePurpose, LocatorPayload, SpecificLocator,
+        TransparentLocator,
+    };
+    use hayagriva_format::io::from_yaml_str;
+
+    #[test]
+    fn test_csl() {
+        let workspace = Path::new(env!("CARGO_MANIFEST_DIR"));
+        let en_locale =
+            fs::read_to_string(workspace.join("tests/data/locales-en-US.xml")).unwrap();
+        let en_locale = LocaleFile::from_xml(&en_locale).unwrap();
+
+        let yaml = fs::read_to_string(workspace.join("tests/data/basic.yml")).unwrap();
+        let bib = from_yaml_str(&yaml).unwrap();
+        let en_locale = [en_locale.into()];
+
+        for style_thing in fs::read_dir(workspace.join("styles/")).unwrap().take(100) {
+            let thing = style_thing.unwrap();
+            if thing.file_type().unwrap().is_dir() {
+                continue;
+            }
+
+            let path = thing.path();
+            let extension = path.extension();
+            if let Some(extension) = extension {
+                if extension.to_str() != Some("csl") {
+                    continue;
+                }
+            } else {
+                continue;
+            }
+
+            println!("testing {path:?}");
+            let style = fs::read_to_string(path).unwrap();
+            let style = IndependentStyle::from_xml(&style).unwrap();
+            let mut driver = BibliographyDriver::new();
+
+            for n in (0..bib.len()).step_by(2) {
+                let items = vec![
+                    CitationItem::with_entry(bib.nth(n).unwrap()),
+                    CitationItem::with_entry(bib.nth(n + 1).unwrap()),
+                ];
+                driver.citation(CitationRequest::new(
+                    items, &style, None, &en_locale, None,
+                ));
+            }
+
+            driver.finish(BibliographyRequest {
+                style: &style,
+                locale: None,
+                locale_files: &en_locale,
+            });
+
+            // for cite in finished.citations {
+            //     println!("{}", cite.citation.to_string(BufWriteFormat::Plain))
+            // }
+
+            // if let Some(bib) = finished.bibliography {
+            //     for (prefix, item) in bib.items {
+            //         if let Some(prefix) = prefix {
+            //             print!("{} ", prefix.to_string(BufWriteFormat::Plain))
+            //         }
+            //         println!("{}", item.to_string(BufWriteFormat::Html))
+            //     }
+            // }
+        }
+    }
+
+    #[test]
+    fn test_transparent_locator_ibid() {
+        let test_data = r#"
+            katalog:
+                type: Book
+                title: "Book"
+                author: "Surname, Name"
+                location: Warsaw
+                date: 1971
+            katalog2:
+                type: Book
+                title: "Book2"
+                author: "Surname, Name"
+                location: Warsaw
+                date: 1971
+        "#;
+        let test_style = r#"<?xml version="1.0" encoding="utf-8"?>
+        <style xmlns="http://purl.org/net/xbiblio/csl" class="note" version="1.0" demote-non-dropping-particle="sort-only">
+           <info>
+              <title>Citing style</title>
+              <id>http://www.example.com/</id>
+           </info>
+
+           <citation>
+              <layout prefix="" suffix="." delimiter="; ">
+                 <choose>
+                    <if position="first">
+                      <text value="first" />
+                    </if>
+                    <else-if position="ibid-with-locator">
+                      <text value="ibid-with-locator" />
+                    </else-if>
+                    <else-if position="ibid">
+                       <text value="ibid" />
+                    </else-if>
+                    <else-if position="subsequent">
+                      <text value="subsequent" />
+                    </else-if>
+                    <else>
+                      <text value="other" />
+                    </else>
+                 </choose>
+              </layout>
+           </citation>
+          <bibliography>
+            <sort>
+              <key macro="contributors-biblio"/>
+            </sort>
+            <layout suffix=".">
+              <group delimiter=", ">
+                <group delimiter=", ">
+                  <text macro="locators-chapter"/>
+                </group>
+              </group>
+            </layout>
+          </bibliography>
+        </style>"#;
+
+        let library = from_yaml_str(test_data).unwrap();
+        let style = citationberg::Style::from_xml(test_style).unwrap();
+        let citationberg::Style::Independent(style) = style else { unreachable!() };
+
+        let mut driver = BibliographyDriver::new();
+        let citations = [
+            ("katalog", Some([2, 3]), "first."),
+            ("katalog", Some([2, 3]), "ibid."),
+            ("katalog", Some([2, 3]), "ibid."),
+            ("katalog", Some([2, 4]), "ibid-with-locator."),
+            ("katalog", None, "subsequent."),
+            ("katalog", Some([2, 3]), "ibid-with-locator."),
+            ("katalog", Some([2, 3]), "ibid."),
+            ("katalog2", Some([2, 3]), "first."),
+        ];
+
+        for (entry, supplement, _) in citations {
+            let entry = library.get(entry).unwrap();
+            driver.citation(CitationRequest::new(
+                vec![CitationItem::with_locator(
+                    entry,
+                    supplement.map(|s| {
+                        SpecificLocator(
+                            Locator::Supplement,
+                            LocatorPayload::Transparent(TransparentLocator(Arc::new(s))),
+                        )
+                    }),
+                )],
+                &style,
+                None,
+                &[],
+                None,
+            ));
+        }
+
+        let finished = driver.finish(BibliographyRequest {
+            style: &style,
+            locale: None,
+            locale_files: &[],
+        });
+
+        for (citation, (_, _, expected_value)) in finished.citations.iter().zip(citations)
+        {
+            let mut s = String::new();
+            citation.citation.write_buf(&mut s, BufWriteFormat::Plain).unwrap();
+            assert_eq!(s, expected_value);
+        }
+    }
+
+    #[test]
+    fn test_chapter_field() {
+        let workspace = Path::new(env!("CARGO_MANIFEST_DIR"));
+        let yaml = fs::read_to_string(workspace.join("tests/data/basic.yml")).unwrap();
+        let bib = from_yaml_str(&yaml).unwrap();
+        let book = bib.get("lamb").unwrap();
+        assert_eq!(book.chapter().unwrap(), &MaybeTyped::Typed(Numeric::new(20)));
+        assert_eq!(book.volume().unwrap(), &MaybeTyped::Typed(Numeric::new(10)));
+        assert_eq!(book.keyed_serial_number("serial"), None);
+        assert_eq!(book.resolve_number_variable(NumberVariable::Number), None);
+
+        let book = bib.get("snail").unwrap();
+        assert_eq!(
+            book.chapter().unwrap(),
+            &MaybeTyped::Typed(Numeric {
+                value: NumericValue::Number(2),
+                prefix: None,
+                suffix: Some(Box::new("A".into()))
+            })
+        );
+        assert_eq!(
+            book.volume().unwrap(),
+            &MaybeTyped::Typed(Numeric {
+                value: NumericValue::Number(4),
+                prefix: None,
+                suffix: Some(Box::new("B".into()))
+            })
+        );
+        assert_eq!(book.keyed_serial_number("serial"), None);
+        assert_eq!(book.resolve_number_variable(NumberVariable::Number), None);
+
+        let chapter = bib.get("lamb-chapter").unwrap();
+        assert_eq!(chapter.chapter().unwrap(), &MaybeTyped::Typed(Numeric::new(3)));
+        assert_eq!(chapter.keyed_serial_number("serial"), None);
+        assert_eq!(
+            chapter.resolve_number_variable(NumberVariable::Number),
+            Some(MaybeTyped::Typed(Cow::Borrowed(&Numeric::new(3))))
+        );
+        assert_eq!(chapter.volume(), None);
+        assert_eq!(
+            chapter.parents().first().unwrap().volume().unwrap(),
+            &MaybeTyped::Typed(Numeric::new(10))
+        );
+
+        let chapter = bib.get("snail-chapter").unwrap();
+        assert_eq!(chapter.chapter(), None);
+        assert_eq!(chapter.keyed_serial_number("serial").unwrap(), "3");
+
+        assert_eq!(chapter.volume(), None);
+        assert_eq!(
+            chapter.parents().first().unwrap().volume().unwrap(),
+            &MaybeTyped::Typed(Numeric {
+                value: NumericValue::Number(4),
+                prefix: None,
+                suffix: Some(Box::new("B".into()))
+            })
+        );
+    }
+
+    #[test]
+    fn test_alphanumeric_disambiguation() {
+        let bibtex = r#"@article{chenTransMorphTransformerUnsupervised2021,
+        title = {{{TransMorph}}: {{Transformer}} for Unsupervised Medical Image Registration},
+        author = {Chen, Junyu and Frey, Eric C. and He, Yufan and Segars, William P. and Li, Ye and Du, Yong},
+        date = {2021},
+        }
+
+        @article{chenViTVNetVisionTransformer2021,
+        title = {{{ViT-V-Net}}: {{Vision Transformer}} for {{Unsupervised Volumetric Medical Image Registration}}},
+        author = {Chen, Junyu and He, Yufan and Frey, Eric C. and Li, Ye and Du, Yong},
+        date = {2021},
+        }"#;
+
+        let library = crate::io::from_biblatex_str(bibtex).unwrap();
+        let alphanumeric = hayagriva_archive::ArchivedStyle::Alphanumeric.get();
+        let citationberg::Style::Independent(alphanumeric) = alphanumeric else {
+            unreachable!()
+        };
+
+        let mut driver = BibliographyDriver::new();
+        for entry in library.iter() {
+            driver.citation(CitationRequest::new(
+                vec![CitationItem::with_entry(entry)],
+                &alphanumeric,
+                None,
+                &[],
+                None,
+            ));
+        }
+
+        let finished = driver.finish(BibliographyRequest {
+            style: &alphanumeric,
+            locale: None,
+            locale_files: &[],
+        });
+
+        let mut c1 = String::new();
+        let mut c2 = String::new();
+
+        finished.citations[0]
+            .citation
+            .write_buf(&mut c1, BufWriteFormat::Plain)
+            .unwrap();
+        finished.citations[1]
+            .citation
+            .write_buf(&mut c2, BufWriteFormat::Plain)
+            .unwrap();
+
+        assert_eq!(c1, "[Che+21a]");
+        assert_eq!(c2, "[Che+21b]");
+    }
+
+    #[test]
+    /// See https://github.com/typst/hayagriva/issues/243
+    fn issue_243() {
+        let bibtex = r#"@book{downs57,
+            title = {An Economic Theory of Democracy},
+            author = {Downs, Anthony},
+            date = {1957},
+            edition = {1},
+            publisher = {Harper \& Row},
+            location = {New York},
+            langid = {english}
+            }
+
+            @book{brady_collier10,
+            title = {Rethinking Social Inquiry. Diverse Tools, Shared Standards},
+            editor = {Brady, Henry E. and Collier, David},
+            date = {2010},
+            edition = {2},
+            publisher = {Rowman \& Littlefield Publishers},
+            location = {Maryland}
+            }"#;
+
+        let library = crate::io::from_biblatex_str(bibtex).unwrap();
+        let apa =
+            hayagriva_archive::ArchivedStyle::AmericanPsychologicalAssociation.get();
+        let citationberg::Style::Independent(apa) = apa else { unreachable!() };
+
+        let locales = hayagriva_archive::locales();
+
+        let mut driver = BibliographyDriver::new();
+        for entry in library.iter() {
+            driver.citation(CitationRequest::new(
+                vec![CitationItem::new(
+                    entry,
+                    None,
+                    Some(LocaleCode::en_us()),
+                    false,
+                    Some(CitePurpose::Prose),
+                )],
+                &apa,
+                Some(LocaleCode::en_us()),
+                &locales,
+                None,
+            ));
+        }
+
+        let finished = driver.finish(BibliographyRequest {
+            style: &apa,
+            locale: Some(LocaleCode::en_us()),
+            locale_files: &locales,
+        });
+
+        let mut c1 = String::new();
+        let mut c2 = String::new();
+
+        finished.citations[0]
+            .citation
+            .write_buf(&mut c1, BufWriteFormat::Plain)
+            .unwrap();
+        finished.citations[1]
+            .citation
+            .write_buf(&mut c2, BufWriteFormat::Plain)
+            .unwrap();
+
+        assert_eq!(c1, "Downs (1957)");
+        assert_eq!(c2, "Brady & Collier (2010)");
+    }
+
+    #[test]
+    /// See https://github.com/typst/hayagriva/issues/48
+    fn issue_48() {
+        let bibtex = r#"@article{chenTransMorphTransformerUnsupervised2021,
+        title = {{{TransMorph}}: {{Transformer}} for Unsupervised Medical Image Registration},
+        author = {Chen, Junyu and Frey, Eric C. and He, Yufan and Segars, William P. and Li, Ye and Du, Yong},
+        date = {2021},
+        }
+
+        @article{chenViTVNetVisionTransformer2021,
+        title = {{{ViT-V-Net}}: {{Vision Transformer}} for {{Unsupervised Volumetric Medical Image Registration}}},
+        author = {Chen, Junyu and He, Yufan and Frey, Eric C. and Li, Ye and Du, Yong},
+        date = {2021},
+        }"#;
+
+        let library = crate::io::from_biblatex_str(bibtex).unwrap();
+        let alphanumeric = hayagriva_archive::ArchivedStyle::Alphanumeric.get();
+        let citationberg::Style::Independent(alphanumeric) = alphanumeric else {
+            unreachable!()
+        };
+
+        let mut driver = BibliographyDriver::new();
+
+        let locator = SpecificLocator(Locator::Custom, LocatorPayload::Str("12"));
+        for entry in library.iter() {
+            driver.citation(CitationRequest::new(
+                vec![CitationItem::with_locator(entry, Some(locator.clone()))],
+                &alphanumeric,
+                None,
+                &[],
+                None,
+            ));
+        }
+
+        driver.citation(CitationRequest::new(
+            vec![CitationItem::with_locator(library.iter().next().unwrap(), None)],
+            &alphanumeric,
+            None,
+            &[],
+            None,
+        ));
+
+        let finished = driver.finish(BibliographyRequest {
+            style: &alphanumeric,
+            locale: None,
+            locale_files: &[],
+        });
+
+        let mut c1 = String::new();
+        let mut c2 = String::new();
+        let mut c3 = String::new();
+
+        finished.citations[0]
+            .citation
+            .write_buf(&mut c1, BufWriteFormat::Plain)
+            .unwrap();
+        finished.citations[1]
+            .citation
+            .write_buf(&mut c2, BufWriteFormat::Plain)
+            .unwrap();
+        finished.citations[2]
+            .citation
+            .write_buf(&mut c3, BufWriteFormat::Plain)
+            .unwrap();
+
+        assert_eq!(c1, "[Che+21a, 12]");
+        assert_eq!(c2, "[Che+21b, 12]");
+        assert_eq!(c3, "[Che+21a]");
+    }
+
+    #[test]
+    fn ibid_handling_with_deutsche_sprache_csl() {
+        let bibtex = r#"@book{ITEM,
+            title = {A},
+            type = {book},
+            }"#;
+
+        let library = crate::io::from_biblatex_str(bibtex).unwrap();
+        let style = hayagriva_archive::ArchivedStyle::DeutscheSprache.get();
+        let citationberg::Style::Independent(style) = style else { unreachable!() };
+
+        let mut driver = BibliographyDriver::new();
+        let entry = library.iter().next().unwrap();
+
+        for locator in ["33", "33", "34"] {
+            driver.citation(CitationRequest::new(
+                vec![CitationItem::new(
+                    entry,
+                    Some(SpecificLocator(Locator::Page, LocatorPayload::Str(locator))),
+                    None,
+                    false,
+                    Some(CitePurpose::Prose),
+                )],
+                &style,
+                None,
+                &[],
+                None,
+            ));
+        }
+
+        let finished = driver.finish(BibliographyRequest {
+            style: &style,
+            locale: None,
+            locale_files: &[],
+        });
+
+        let actual = finished
+            .citations
+            .iter()
+            .map(|c| {
+                let mut s = String::new();
+                c.citation.write_buf(&mut s, BufWriteFormat::Plain).unwrap();
+                s
+            })
+            .collect::<Vec<_>>();
+
+        assert_eq!(actual, ["(33)", "()", "(34)"]);
+    }
 }
