+++ crates/hayagriva-format/Cargo.toml
@@ -19,5 +19,13 @@ unic-langid = { workspace = true }
 serde_yaml = "0.9.25"
 unicode-segmentation = { workspace = true }
 hayagriva-core = { path = "../hayagriva-core" }
+hayagriva-biblatex = { path = "../hayagriva-biblatex", optional = true }
 thiserror = { workspace = true }
 unscanny = { workspace = true }
+url = { workspace = true, optional = true }
+
+[dev-dependencies]
+serde_json = "1"
+
+[features]
+biblatex = ["dep:biblatex", "dep:hayagriva-biblatex", "dep:url"]
